+++ src/Spinner/Core/Spinner.php
@@ -3,11 +3,8 @@
 namespace AlecRabbit\Spinner\Core;
 
 use AlecRabbit\Cli\Tools\Cursor;
-use AlecRabbit\Spinner\Core\Adapters\EchoOutputAdapter;
-use AlecRabbit\Spinner\Core\Contracts\Frames;
 use AlecRabbit\Spinner\Core\Contracts\SpinnerInterface;
 use AlecRabbit\Spinner\Core\Contracts\SpinnerOutputInterface;
-use AlecRabbit\Spinner\Core\Contracts\StylesInterface;
 use AlecRabbit\Spinner\Core\Jugglers\Contracts\JugglerInterface;
 use AlecRabbit\Spinner\Core\Jugglers\FrameJuggler;
 use AlecRabbit\Spinner\Core\Jugglers\MessageJuggler;
@@ -17,16 +14,8 @@ use AlecRabbit\Spinner\Settings\Settings;
 use function AlecRabbit\typeOf;
 use const AlecRabbit\ESC;
 
-abstract class Spinner implements SpinnerInterface
+abstract class Spinner extends SpinnerCore
 {
-    protected const EMPTY_STRING = Defaults::EMPTY_STRING;
-
-    protected const INTERVAL = Defaults::DEFAULT_INTERVAL;
-    protected const FRAMES = Frames::BASE;
-    protected const STYLES = StylesInterface::STYLING_DISABLED;
-
-    /** @var null|SpinnerOutputInterface */
-    protected $output;
     /** @var Settings */
     protected $settings;
     /** @var bool */
@@ -81,34 +70,6 @@ abstract class Spinner implements SpinnerInterface
         ];
     }
 
-    /**
-     * @param null|false|SpinnerOutputInterface $output
-     * @return null|SpinnerOutputInterface
-     */
-    protected function refineOutput($output): ?SpinnerOutputInterface
-    {
-        $this->assertOutput($output);
-        if (false === $output) {
-            return null;
-        }
-        return $output ?? new EchoOutputAdapter();
-    }
-
-    /**
-     * @param mixed $output
-     */
-    protected function assertOutput($output): void
-    {
-        if (null !== $output && false !== $output && !$output instanceof SpinnerOutputInterface) {
-            $typeOrValue =
-                true === $output ? 'true' : typeOf($output);
-            throw new \InvalidArgumentException(
-                'Incorrect parameter: ' .
-                '[null|false|' . SpinnerOutputInterface::class . '] expected'
-                . ' "' . $typeOrValue . '" given.'
-            );
-        }
-    }
 
     /**
      * @param null|string|Settings $settings