+++ src/Spinner/Core/SpinnerCore.php
@@ -1,0 +1,52 @@
+<?php declare(strict_types=1);
+
+namespace AlecRabbit\Spinner\Core;
+
+use AlecRabbit\Spinner\Core\Adapters\EchoOutputAdapter;
+use AlecRabbit\Spinner\Core\Contracts\Frames;
+use AlecRabbit\Spinner\Core\Contracts\SpinnerInterface;
+use AlecRabbit\Spinner\Core\Contracts\SpinnerOutputInterface;
+use AlecRabbit\Spinner\Core\Contracts\StylesInterface;
+use AlecRabbit\Spinner\Settings\Contracts\Defaults;
+use function AlecRabbit\typeOf;
+
+abstract class SpinnerCore implements SpinnerInterface
+{
+    protected const EMPTY_STRING = Defaults::EMPTY_STRING;
+
+    protected const INTERVAL = Defaults::DEFAULT_INTERVAL;
+    protected const FRAMES = Frames::BASE;
+    protected const STYLES = StylesInterface::STYLING_DISABLED;
+
+    /** @var null|SpinnerOutputInterface */
+    protected $output;
+
+    /**
+     * @param null|false|SpinnerOutputInterface $output
+     * @return null|SpinnerOutputInterface
+     */
+    protected function refineOutput($output): ?SpinnerOutputInterface
+    {
+        $this->assertOutput($output);
+        if (false === $output) {
+            return null;
+        }
+        return $output ?? new EchoOutputAdapter();
+    }
+
+    /**
+     * @param mixed $output
+     */
+    protected function assertOutput($output): void
+    {
+        if (null !== $output && false !== $output && !$output instanceof SpinnerOutputInterface) {
+            $typeOrValue =
+                true === $output ? 'true' : typeOf($output);
+            throw new \InvalidArgumentException(
+                'Incorrect parameter: ' .
+                '[null|false|' . SpinnerOutputInterface::class . '] expected'
+                . ' "' . $typeOrValue . '" given.'
+            );
+        }
+    }
+}