+++ src/Spinner/LineSpinner.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace AlecRabbit\Spinner;
-
-class LineSpinner extends SimpleSpinner
-{
-}